+++ client/CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `TrackedParticipation::answers` field;
 - `#[derive(Debug, Clone, Serialize, Deserialize)]` and `#[serde(untagged)]` to `ParticipationEventType`;
+- `impl From<&OutputMetadata> for OutputMetadataResponse`;
 
 ### Changed